+++ test/DateTest.php
@@ -2,6 +2,7 @@
 
 use MadisonSolutions\JustDate\DateRange;
 use MadisonSolutions\JustDate\JustDate;
+use MadisonSolutions\JustDate\JustTime;
 use PHPUnit\Framework\TestCase;
 
 class DateTest extends TestCase
@@ -31,7 +32,7 @@ class DateTest extends TestCase
 
     public function testCreateJustDates()
     {
-        $d = new JustDate(2019, 4, 21);
+        $d = JustDate::make(2019, 4, 21);
         $this->assertJustDate('2019-04-21', $d);
     }
 
@@ -52,7 +53,7 @@ class DateTest extends TestCase
         $p2 = new DateTime('2019-04-21 16:23:12', new DateTimeZone('Asia/Calcutta'));
         $this->assertNotEquals($p1->getTimestamp(), $p2->getTimestamp());
 
-        // ...however it should will product an equal JustDate object
+        // ...however it should produce the same JustDate object
         $d2 = JustDate::fromDateTime($p2);
         $this->assertJustDate('2019-04-21', $d2);
         $this->assertSame($d1->timestamp, $d2->timestamp);
@@ -118,17 +119,28 @@ class DateTest extends TestCase
         });
     }
 
-    public function testSpanDays()
+    public function testEpochDay()
     {
-        $this->assertSame(2, JustDate::spanDays(new JustDate(2019, 04, 21), new JustDate(2019, 04, 23)));
-        $this->assertSame(0, JustDate::spanDays(new JustDate(2019, 04, 21), new JustDate(2019, 04, 21)));
-        $this->assertSame(-2, JustDate::spanDays(new JustDate(2019, 04, 23), new JustDate(2019, 04, 21)));
-        $this->assertSame(365, JustDate::spanDays(new JustDate(2018, 04, 21), new JustDate(2019, 04, 21)));
+        $this->assertJustDate('1970-01-01', JustDate::fromEpochDay(0));
+        $this->assertJustDate('1970-01-02', JustDate::fromEpochDay(1));
+        $this->assertJustDate('1969-12-31', JustDate::fromEpochDay(-1));
+        $this->assertEquals(0, JustDate::fromYmd('1970-01-01')->epoch_day);
+        $this->assertEquals(1, JustDate::fromYmd('1970-01-02')->epoch_day);
+        $this->assertEquals(-1, JustDate::fromYmd('1969-12-31')->epoch_day);
+        $this->assertEquals(strtotime('2021-04-28T00:00:00.03+00:00') / (60 * 60 * 24), JustDate::fromYmd('2021-04-28')->epoch_day);
+    }
+
+    public function testNumNights()
+    {
+        $this->assertSame(2, JustDate::numNights(JustDate::make(2019, 04, 21), JustDate::make(2019, 04, 23)));
+        $this->assertSame(0, JustDate::numNights(JustDate::make(2019, 04, 21), JustDate::make(2019, 04, 21)));
+        $this->assertSame(-2, JustDate::numNights(JustDate::make(2019, 04, 23), JustDate::make(2019, 04, 21)));
+        $this->assertSame(365, JustDate::numNights(JustDate::make(2018, 04, 21), JustDate::make(2019, 04, 21)));
     }
 
     public function testGetters()
     {
-        $d = new JustDate(2019, 04, 21);
+        $d = JustDate::make(2019, 04, 21);
         $this->assertSame(2019, $d->year);
         $this->assertSame(4, $d->month);
         $this->assertSame(21, $d->day);
@@ -138,7 +150,7 @@ class DateTest extends TestCase
 
     public function testAddDays()
     {
-        $d1 = new JustDate(2019, 04, 21);
+        $d1 = JustDate::make(2019, 04, 21);
         $this->assertJustDate('2019-04-22', $d1->addDays(1));
         $this->assertJustDate('2019-04-22', $d1->nextDay());
         $this->assertJustDate('2019-05-01', $d1->addDays(10));
@@ -166,11 +178,40 @@ class DateTest extends TestCase
         // Check some edge cases
         // Adding one month to Jan 30 is ambiguous because there is no Feb 30
         // Expect it to overflow to March 1/2 (depending on if it's a leap year)
-        $this->assertJustDate('2020-03-01', (new JustDate(2020, 01, 30))->addMonths(1));
-        $this->assertJustDate('2021-03-02', (new JustDate(2021, 01, 30))->addMonths(1));
-        $this->assertJustDate('2021-03-30', (new JustDate(2021, 01, 30))->addMonths(2));
-        $this->assertJustDate('2021-03-02', (new JustDate(2021, 03, 30))->addMonths(-1));
+        $this->assertJustDate('2020-03-01', (JustDate::make(2020, 01, 30))->addMonths(1));
+        $this->assertJustDate('2021-03-02', (JustDate::make(2021, 01, 30))->addMonths(1));
+        $this->assertJustDate('2021-03-30', (JustDate::make(2021, 01, 30))->addMonths(2));
+        $this->assertJustDate('2021-03-02', (JustDate::make(2021, 03, 30))->addMonths(-1));
         // Note this means sometimes $d->addMonths(a)->addMonths(b) is not equal to $d->addMonths(a + b) !
+    }
+
+    public function testSubDays()
+    {
+        $d1 = JustDate::make(2019, 04, 21);
+        $this->assertJustDate('2019-04-20', $d1->subDays(1));
+        $this->assertJustDate('2019-04-11', $d1->subDays(10));
+        $this->assertJustDate('2018-06-25', $d1->subDays(300));
+        $this->assertJustDate('2019-04-22', $d1->subDays(-1));
+        $this->assertJustDate('2020-02-15', $d1->subDays(-300));
+
+        $this->assertJustDate('2019-04-14', $d1->subWeeks(1));
+        $this->assertJustDate('2019-03-31', $d1->subWeeks(3));
+        $this->assertJustDate('2019-04-28', $d1->subWeeks(-1));
+
+        $this->assertJustDate('2019-03-21', $d1->subMonths(1));
+        $this->assertJustDate('2019-02-21', $d1->subMonths(2));
+        $this->assertJustDate('2019-05-21', $d1->subMonths(-1));
+
+        $this->assertJustDate('2018-04-21', $d1->subYears(1));
+        $this->assertJustDate('2017-04-21', $d1->subYears(2));
+        $this->assertJustDate('2020-04-21', $d1->subYears(-1));
+
+        // Check some edge cases
+        // Subtracting one month from Mar 30 is ambiguous because there is no Feb 30
+        // Expect it to overflow to March 1/2 (depending on if it's a leap year)
+        $this->assertJustDate('2020-03-01', (JustDate::make(2020, 03, 30))->subMonths(1));
+        $this->assertJustDate('2021-03-02', (JustDate::make(2021, 03, 30))->subMonths(1));
+        $this->assertJustDate('2021-01-30', (JustDate::make(2021, 03, 30))->subMonths(2));
     }
 
     public function testFormat()
@@ -188,8 +229,8 @@ class DateTest extends TestCase
 
     public function testComparisons()
     {
-        $d1 = new JustDate(2019, 04, 21);
-        $d2 = new JustDate(2019, 04, 22);
+        $d1 = JustDate::make(2019, 04, 21);
+        $d2 = JustDate::make(2019, 04, 22);
 
         $this->assertTrue($d1->isBefore($d2));
         $this->assertFalse($d2->isBefore($d1));
@@ -207,40 +248,40 @@ class DateTest extends TestCase
 
     public function testDateTrickery()
     {
-        $d1 = new JustDate(2019, 04, 0); // day = 0 gives last day of prev month
+        $d1 = JustDate::make(2019, 04, 0); // day = 0 gives last day of prev month
         $this->assertJustDate('2019-03-31', $d1);
 
-        $d2 = new JustDate(2019, 1, 0);
+        $d2 = JustDate::make(2019, 1, 0);
         $this->assertJustDate('2018-12-31', $d2);
     }
 
     public function testStartAndEndOfMonths()
     {
-        $d1 = new JustDate(2019, 04, 21);
+        $d1 = JustDate::make(2019, 04, 21);
         $this->assertJustDate('2019-04-01', $d1->startOfMonth());
         $this->assertJustDate('2019-04-30', $d1->endOfMonth());
     }
 
     public function testEarliestAndLatest()
     {
-        $d1 = new JustDate(2019, 04, 21);
+        $d1 = JustDate::make(2019, 04, 21);
         $this->assertJustDate('2019-04-21', JustDate::earliest($d1));
         $this->assertJustDate('2019-04-21', JustDate::latest($d1));
         $this->assertJustDate('2019-04-21', JustDate::earliest($d1, $d1));
         $this->assertJustDate('2019-04-21', JustDate::latest($d1, $d1));
 
-        $d2 = new JustDate(2019, 04, 22);
+        $d2 = JustDate::make(2019, 04, 22);
         $this->assertJustDate('2019-04-21', JustDate::earliest($d1, $d2));
         $this->assertJustDate('2019-04-22', JustDate::latest($d1, $d2));
 
-        $d3 = new JustDate(2019, 04, 23);
+        $d3 = JustDate::make(2019, 04, 23);
         $this->assertJustDate('2019-04-21', JustDate::earliest($d3, $d2, $d1));
         $this->assertJustDate('2019-04-23', JustDate::latest($d3, $d2, $d1));
     }
 
     public function testDaysOfTheWeek()
     {
-        $d0 = new JustDate(2021, 03, 01); // Monday 1st March
+        $d0 = JustDate::make(2021, 03, 01); // Monday 1st March
         for ($i = 0; $i < 7; $i++) {
             $d = $d0->addDays($i);
             $is_days = [
@@ -269,10 +310,38 @@ class DateTest extends TestCase
         }
     }
 
+    public function testConversionToDateTime()
+    {
+        $default_timezone = new DateTimeZone(date_default_timezone_get());
+        $tahiti = new DateTimeZone('Pacific/Tahiti');
+
+        $d1 = JustDate::make(2021, 04, 28);
+
+        // With no parameters, we should get midnight, in the system default timezone
+        $td = $d1->toDateTime();
+        $this->assertEquals('2021-04-28 00:00:00', $td->format('Y-m-d H:i:s'));
+        $this->assertEquals($default_timezone, $td->getTimezone());
+
+        // Try specifying a time
+        $td = $d1->toDateTime(JustTime::fromHis('14:35:02'));
+        $this->assertEquals('2021-04-28 14:35:02', $td->format('Y-m-d H:i:s'));
+        $this->assertEquals($default_timezone, $td->getTimezone());
+
+        // Try specifying a timezone
+        $td = $d1->toDateTime(null, $tahiti);
+        $this->assertEquals('2021-04-28 00:00:00', $td->format('Y-m-d H:i:s'));
+        $this->assertEquals($tahiti, $td->getTimezone());
+
+        // Try specifying both
+        $td = $d1->toDateTime(JustTime::fromHis('14:35:02'), $tahiti);
+        $this->assertEquals('2021-04-28 14:35:02', $td->format('Y-m-d H:i:s'));
+        $this->assertEquals($tahiti, $td->getTimezone());
+    }
+
     public function testCreateRange()
     {
-        $d1 = new JustDate(2019, 04, 21);
-        $d2 = new JustDate(2019, 04, 25);
+        $d1 = JustDate::make(2019, 04, 21);
+        $d2 = JustDate::make(2019, 04, 25);
         $r = new DateRange($d1, $d2);
         $this->assertDateRange("2019-04-21 to 2019-04-25", $r);
 
@@ -281,7 +350,7 @@ class DateTest extends TestCase
 
         // Won't let you make a range with end before start
         $this->assertThrows(InvalidArgumentException::class, function () {
-            new DateRange(new JustDate(2019, 04, 21), new JustDate(2019, 04, 19));
+            new DateRange(JustDate::make(2019, 04, 21), JustDate::make(2019, 04, 19));
         });
         $this->assertThrows(InvalidArgumentException::class, function () {
             DateRange::fromYmd('2019-04-21', '2019-04-19');
@@ -305,14 +374,12 @@ class DateTest extends TestCase
         $r1 = DateRange::fromYmd('2019-04-21', '2019-04-25');
         $this->assertJustDate('2019-04-21', $r1->start);
         $this->assertJustDate('2019-04-25', $r1->end);
-        $this->assertSame(4, $r1->span);
         $this->assertSame(4, $r1->num_nights);
         $this->assertSame(5, $r1->num_days);
 
         $r2 = DateRange::fromYmd('2019-04-21', '2019-04-21');
         $this->assertJustDate('2019-04-21', $r2->start);
         $this->assertJustDate('2019-04-21', $r2->end);
-        $this->assertSame(0, $r2->span);
         $this->assertSame(0, $r2->num_nights);
         $this->assertSame(1, $r2->num_days);
     }
@@ -320,11 +387,11 @@ class DateTest extends TestCase
     public function testRangeIncludes()
     {
         $r1 = DateRange::fromYmd('2019-04-21', '2019-04-25');
-        $this->assertFalse($r1->includes(new JustDate(2019, 04, 20)));
-        $this->assertTrue($r1->includes(new JustDate(2019, 04, 21)));
-        $this->assertTrue($r1->includes(new JustDate(2019, 04, 23)));
-        $this->assertTrue($r1->includes(new JustDate(2019, 04, 25)));
-        $this->assertFalse($r1->includes(new JustDate(2019, 04, 27)));
+        $this->assertFalse($r1->includes(JustDate::make(2019, 04, 20)));
+        $this->assertTrue($r1->includes(JustDate::make(2019, 04, 21)));
+        $this->assertTrue($r1->includes(JustDate::make(2019, 04, 23)));
+        $this->assertTrue($r1->includes(JustDate::make(2019, 04, 25)));
+        $this->assertFalse($r1->includes(JustDate::make(2019, 04, 27)));
     }
 
     public function testRangeIterators()
@@ -452,14 +519,14 @@ class DateTest extends TestCase
 
     public function testSerialization()
     {
-        $d1 = new JustDate(2019, 04, 21);
+        $d1 = JustDate::make(2019, 04, 21);
         $s = serialize($d1);
         $this->assertTrue(is_string($s));
         $_d1 = unserialize($s);
         $this->assertJustDate('2019-04-21', $_d1);
         $this->assertNotSame($d1, $_d1);
 
-        $d2 = new JustDate(2019, 04, 25);
+        $d2 = JustDate::make(2019, 04, 25);
         $r = new DateRange($d1, $d2);
         $s = serialize($r);
         $this->assertTrue(is_string($s));